refactor: simplier git module import
