+++ lib/jenkins.js
@@ -1,4 +1,4 @@
-const git = require('../lib/git');
+const git = require('./git');
 
 // https://wiki.jenkins.io/display/JENKINS/Building+a+software+project
