+++ lib/semaphore.js
@@ -1,4 +1,4 @@
-const git = require('../lib/git');
+const git = require('./git');
 
 // https://semaphoreci.com/docs/available-environment-variables.html
